LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/historical-expected.txt


FAIL Historical CSS features must be removed: WebKitTransitionEvent assert_equals: expected (undefined) undefined but got (function) function "function WebKitTransitionEvent() {
    [native code]
}"
